requirements.txt
Django==1.8.0
MySQL-python==1.2.5
stripe==1.22.2
wheel==0.24.0
Pillow==2.6.1
boto==2.38.0
django-storages==1.1.8
django-nested-inline=0.3.5
